requirements.txt
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-MySQLdb==1.0.1
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
protobuf==3.20.3
Werkzeug==2.2.3
